django-farout/requirements.txt
# Core Django
Django==5.1.3
python-decouple==3.8
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Authentication
django-allauth[socialaccount]==65.0.2

# REST API (future-ready)
djangorestframework==3.15.2

# Security
django-csp==3.8

# Admin & Forms
django-crispy-forms==2.3
crispy-tailwind==1.0.3

# Rich Text Editor
django-tinymce==4.1.0

# Production Server
gunicorn==22.0.0
whitenoise==6.7.0

# Development
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Utilities
python-slugify==8.0.4
Pillow==10.4.0
requests==2.32.3
